SHCS/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
PyJWT==2.8.0
redis==5.0.1
bcrypt==4.0.1
python-dotenv==1.0.0
PyMySQL==1.1.0
Werkzeug==2.3.7
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.26.1
eventlet==0.33.3
cryptography==41.0.4
Flask-Limiter==3.5.0
requests==2.31.0 